Created fix for bug #5
kmt now checks for the name and version parameters before trying to artifact a manifest. In my experience this have been things that are not being versioned the same way as applications and are infrastructure manifests.

These are not updated as frequently
